Refactoring the wait_for_semaphore function to utils.py
